estrazione_consumi/requirements.txt
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
beautifulsoup4==4.12.3
pandas==2.2.1
python-dateutil==2.9.0
lxml==5.1.0
html5lib==1.1
openpyxl==3.1.2
